Python/Socket/Smart_Home/requirements.txt
grpcio_tools
kafka-python
